+++ lib/entity/vo/image_item_vo.dart
@@ -2,6 +2,8 @@ class ImageItemVO {
 
   const ImageItemVO(this.id, this.name, this.avatarUrl, this.imageUrl);
 
+  const ImageItemVO.empty() : this(0, "", "", "");
+
   final int id;
   final String name;
   final String avatarUrl;