+++ lib/logic/picture_paging_logic.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:dio/dio.dart';
 import 'package:flutter_pixabay/repository/picture_repository.dart';
 
@@ -7,7 +9,7 @@ class PicturePagingLogic {
   PicturePagingLogic(this._repo);
 
   final PictureRepository _repo;
-  ImageQueryIntent? _queryIntent;
+  _ImageQueryIntent? _queryIntent;
 
   Future<PixabayRemoteData?> searchImage(String query) async {
     var intent = _newQueryIntent(query);
@@ -18,15 +20,15 @@ class PicturePagingLogic {
     return await _queryIntent?.fetchNextPageData();
   }
 
-  ImageQueryIntent _newQueryIntent(String query) {
+  _ImageQueryIntent _newQueryIntent(String query) {
     _queryIntent?.cancel();
-    var newIntent = ImageQueryIntent(query, _repo);
+    var newIntent = _ImageQueryIntent(query, _repo);
     _queryIntent = newIntent;
     return newIntent;
   }
 }
 
-class ImageQueryIntent {
+class _ImageQueryIntent {
   final String _query;
   final PictureRepository _repo;
 
@@ -34,7 +36,9 @@ class ImageQueryIntent {
   bool _isFetching = false;
   CancelToken? _cancelToken;
 
-  ImageQueryIntent(this._query, this._repo);
+  final StreamController _streamController = StreamController();
+
+  _ImageQueryIntent(this._query, this._repo);
 
   Future<PixabayRemoteData?> fetchNextPageData() async {
     if(_isFetching) return null;
@@ -42,6 +46,7 @@ class ImageQueryIntent {
 
     _cancelToken = CancelToken();
     var searchResponse = await _repo.searchImage(_query, page: _currentPage + 1, cancelToken: _cancelToken);
+    _streamController.add(searchResponse.hits);
     _currentPage += 1;
     _isFetching = false;
     return searchResponse;