+++ changelog.md
@@ -1,3 +1,15 @@
+[2023-04-27]
+- [Refactor Home] Use bloc provider.
+
+[2023-04-25]
+- [Refactor Home] Use bloc pattern to re-write logic.
+
+[2023-04-20]
+- Refactor home page
+  - [Router] Refactor router
+  - [DI] home page dependency refactor
+  - [UI] 拆分元件
+  
 [2023-04-19]
 - Add loading indicator
   - [Home UI] Change `opacity` when loading progressing.